chore: migrate to use `diagram-js-ui`
