+++ package.json
@@ -29,10 +29,12 @@
   },
   "license": "MIT",
   "dependencies": {
+    "@bpmn-io/diagram-js-ui": "^0.2.2",
     "bpmn-moddle": "^8.0.0",
     "bpmnlint": "^9.2.0",
     "bpmnlint-plugin-camunda-compat": "^2.1.0",
     "bpmnlint-utils": "^1.0.2",
+    "clsx": "^2.0.0",
     "min-dash": "^4.0.0",
     "min-dom": "^4.1.0",
     "modeler-moddle": "^0.2.0",
